+++ core/internal/features/ocr2/features_ocr2_test.go
@@ -85,6 +85,7 @@ func setupOCR2Contracts(t *testing.T) (*bind.TransactOpts, *backends.SimulatedBa
 		accessAddress,
 		9,
 		"TEST",
+		false,
 	)
 	// Ensure we have finality depth worth of blocks to start.
 	for i := 0; i < 20; i++ {