+++ core/services/relay/evm/evm.go
@@ -291,13 +291,16 @@ func newConfigProvider(lggr logger.Logger, chain evm.Chain, opts *types.RelayOpt
 	if relayConfig.FeedID != nil {
 		cp, err = mercury.NewConfigPoller(
 			lggr,
+			chain.Client(),
 			chain.LogPoller(),
 			contractAddress,
 			*relayConfig.FeedID,
 			eventBroadcaster,
 		)
 	} else {
-		cp, err = NewConfigPoller(lggr,
+		cp, err = NewConfigPoller(
+			lggr,
+			chain.Client(),
 			chain.LogPoller(),
 			contractAddress,
 		)