+++ core/services/relay/evm/functions/config_poller_test.go
@@ -68,6 +68,7 @@ func runTest(t *testing.T, pluginType functions.FunctionsPluginType, expectedDig
 		accessAddress,
 		9,
 		"TEST",
+		false,
 	)
 	require.NoError(t, err)
 	b.Commit()