+++ core/services/relay/evm/types/types.go
@@ -5,9 +5,15 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"math/big"
+	"time"
 
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/jpillora/backoff"
 	"github.com/lib/pq"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 
 	"gopkg.in/guregu/null.v2"
 
@@ -18,6 +24,29 @@ import (
 
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
+
+var (
+	FailedRPCContractCalls = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "ocr2_failed_rpc_contract_calls",
+		Help: "Running count of failed RPC contract calls to OCR2 configuration contract",
+	},
+		[]string{"chainID", "contractAddress", "feedID"},
+	)
+)
+
+func NewRPCCallBackoff() backoff.Backoff {
+	return backoff.Backoff{
+		Factor: 2,
+		Jitter: true,
+		Min:    100 * time.Millisecond,
+		Max:    1 * time.Hour,
+	}
+}
+
+type ContractCaller interface {
+	CallContract(ctx context.Context, msg ethereum.CallMsg, blockNumber *big.Int) ([]byte, error)
+	ConfiguredChainID() *big.Int
+}
 
 type RelayConfig struct {
 	ChainID                *utils.Big  `json:"chainID"`